Move Table of Content code to separate directory
