+++ src_js/components/sidebar/table_of_contents/unflattenHeadings.ts
